#3 Implemented IFolderCreationService
Refactored FolderDeletionService --> FolderServices and implemented
IFolderCreationService.
